U-shape test for rare events logistic regression

U-shape test for rare events logistic regression

Usage

U_shape_test(
  data,
  dep,
  ind,
  cnt,
  mod = "brglm",
  boot = c("none", "nonparam", "factor"),
  factor = NULL,
  n = 1000,
  HC = T,
  const = T,
  table_remove = NULL,
  plot = F,
  tab_save = F
)

Arguments

data

Dataframe

dep

Character. Dependent variable.

ind

Character. Independent variable.

cnt

Vector of characters with names of control variables.

mod

Character. Type of model (now just brglm is supported)

boot

Character. Should middle point and "lines" be calculated via bootstrap? If no - "none", if yes - one can choose bootstrap type

factor

Character. Only if boot = "factor".

n

Number of bootstrap draws

HC

Logical. Should heteroscedasticity-consistent SE be used?

const

Logical. Should constant be included in the models?

table_remove

Vector of variables names that should be excluded from the regression table.

plot

Logical. Should plot be depicted?

tab_save

Logical. Should table be saved? If yes - table is saved in working directory.

Value

Plot and regression table with u-shape test summary.

References

Lind, J. T., & Mehlum, H. (2010). With or without U? The appropriate test for a U‐shaped relationship. Oxford bulletin of economics and statistics, 72(1), 109-118.

Simonsohn, U. (2018). Two lines: A valid alternative to the invalid testing of U-shaped relationships with quadratic regressions. Advances in Methods and Practices in Psychological Science, 1(4), 538-555.