FGLS model for estimated dependent variable

FGLS model for estimated dependent variable

Usage

edv.fgls(ols, DVvariance = NULL, n = NULL)

Arguments

ols

Initial OLS model (lm in R)

DVvariance

Numeric vector of variances of each y. The length should be equal the number of observations in the initial model.

n

Numeric vector of number of observations (n) on which each y was estimated. The length should be equal the number of observations in the initial model.

Value

re-estimated model

Details

The function is based completely on the following paper: Lewis, J. B., & Linzer, D. A. (2005). Estimating regression models in which the dependent variable is based on estimates. Political analysis, 13(4), 345-364. There are two possible strategies. If variance of each y is unknown, then Numeric vector of number of observations (n) on which each y was estimated should be provided. For example, y - mean of housholds income, then n is the number of housholds. The second possible situation is when a variance of each y is known. In that case, you should provide DVvariance arguement in the function with numeric vector of such variances (as in case with the first situation with number of observations)