Revolutions dataset, 2000-2022 (Version 9)

The database on revolutionary events of the 21st century was prepared within the framework of the project “Quantitative Analysis and Forecasting of the Risks of Socio-Political Destabilization in the Countries of the Afrasian Macrozone of Instability” supported by the Russian Science Foundation (project No. 18-18-00254-P). The database describes revolutionary events of the 21st century by various characteristics: chronological, geographical, type of protestors' tactics, purpose and degree of success.

Usage

revolutions

Format

revolutions A data frame with event as unit of analysis.

n

unique episode number

start_year

event start year

end_year

event end year

country

country name

iso3c

three-letter ISO 3166-1 alpha-3 (iso3c) country code

region

the region to which the country where the event took place belongs (UN Subregion classification)

successful

complete success (1 = yes, 0 = no)

limited

limited success (1 = yes, 0 = no)

failed

fail (1 = yes, 0 = no)

ongoing

is episode still not ended? (1 = yes, 0 = no). Note, episode can be ongoing and has "success" coding that is preliminary estimate.

armed

armed tactic: fabric weapons (1 = yes, 0 = no)

unarmed

unarmed tactic: nonviolent resistance or improvised weapons - sticks, stones (1 = yes, 0 = no)

regime_change

the goal was a regime change (1 = yes, 0 = no)

islamistic

the goal was connected to Islam (1 = yes, 0 = no). Note, this feature usually goes with regime change goal

democratic

the goal was to establish/improve democratic institutions (1 = yes, 0 = no). Note, this feature usually goes with regime change goal

social

the goal was primary social, firstly connected with life-conditions (1 = yes, 0 = no)

separatism

the goal was to gain independence/autonomy (1 = yes, 0 = no)

other

other goal (1 = yes, 0 = no). Can be overlapped by mentioned goal types

coupvolution

Was an event ended with "End game Coup"? (1 = yes, 0 = no)

revolutions

Is event a pure revolution (mass mobilization, aim to overthrough regime)? (1 = yes, 0 = no)

revolutions_plus_question

If event is not a pure revolution, can it be still considered as revolutionary movement? (1 = yes, 0 = no)

all

All types of event including quasi-revolutionary events (one unique value = 1)

ethnic

Was an episode connected to ethnic cleaveges? (1 = yes, 0 = no)

Source

Center for Stability and Risk Analysis